test(js): move #295 bundle-export test to an integration file
Codacy flags std::env::temp_dir() in inline #[cfg(test)] mods (test files are exempt, src/ mods are
not). Move bundle_emits_named_exports to crates/tish_compile_js/tests/bundle_exports.rs using
CARGO_TARGET_TMPDIR (a per-test-binary scratch dir under target/) instead.
